Updated heuristics
